Refactor the form and the top menu from "Edit recipe" page into 2 components. (#38)
* Refactor the form from "Edit recipe" page into EditRecipeForm component.
* Refactor the top menu from "Edit recipe" page into EditRecipeTopMenu component.
